use ConfigurationPropertiesBindingPostProcessor directly
finally fixed...
